backend/app/ai/requirements.txt
langchain==0.1.17
langchain-google-genai
pymupdf==1.24.1
faiss-cpu
tqdm
python-dotenv
openai